+++ src/VehicleManager.Application/Vehicles/Commands/CreateVehicle/CreateVehicleCommand.cs
@@ -8,7 +8,7 @@ internal record CreateVehicleCommand(
     int Year,
     string LicensePlate,
     string Vin,
-    double EngineCapacity,
+    int EngineCapacity,
     int EnginePower,
     FuelType FuelType,
     GearboxType GearboxType,
